Teensy and Adafruit Pro Trinket - Neopixel and FSR boards
Small boards (more like hats) for the two different boards.  There are a few
designs setup to be used with HROS1 as the LED hands.

Likewise the FSR boards are setup to hopefully work in the feet of the
HROS1 to control 4 FSR sensors to detect how the foot is hitting the ground.

warning these are works in progress and have not been fully tested out